// src/app/modules/user-management/users/models/user.model.ts
export interface User {
    id: any;
    firstName: any;
    lastName: any;
    dateOfBbirth: any;
    userName: any;
    gender: any;
    ipAddress: any;
    type: any;
    dob: Date;
    isDeleted: false,
    _id: number,
    createdBy: string,
    role: string,
    email: string,
    name: string
    password: string,
    createdOn: string,
    __v: number,
    lastLogin: string,
    mobileNumber: number
}